remove auth from unit tests
